Solution cs: comment code change needed
within the solution cs file, we need to move the costs to each action, so we will need to break out the action object creation so that each can have the costs attached, which eventually should be refactored to have the cost as part of the constructor?  Certainly if/when it goes to a database.